[Algorithms] docs: Add another standard solution for the problem "Power Set LCCI"
