--- simi-app/simi-memory-management/src/main/resources/temp.txt
@@ -1,5 +1,0 @@
-static final class Node implements ForkJoinPool.ManagedBlocker {
-    final boolean isData;   // false if this is a request node
-    volatile Object item;   // initially non-null if isData; CASed to match
-    volatile Node next;
-    volatile Thread waiter; // null when not waiting for a match
