+++ simi-app/simi-review-tool/pom.xml
@@ -9,7 +9,7 @@
         <version>1.1.0</version>
     </parent>
 
-    <artifactId>simi-memory-management</artifactId>
+    <artifactId>simi-review-tool</artifactId>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
