+++ simi-app/simi-review-tool/src/main/resources/storage.txt
@@ -1,0 +1,29 @@
+Algorithms Problems / Array / Array Partition
+Algorithms Problems / Conditional Logic / Add Edges to Make Degrees of All Nodes Even
+Algorithms Problems / Depth-first Search / Amount of Time for Binary Tree to Be Infected
+Algorithms Problems / Dichotomy / Search in Rotated Sorted Array
+Algorithms Problems / Difference Array / Jump Game VII
+Algorithms Problems / Dynamic Programming / Climbing Stairs
+Algorithms Problems / Dynamic Programming / Count All Valid Pickup and Delivery Options
+Algorithms Problems / Dynamic Programming / Decode Ways II
+Algorithms Problems / Dynamic Programming / Make the XOR of All Segments Equal to Zero
+Algorithms Problems / Dynamic Programming / Maximize Value of Function in a Ball Passing Game
+Algorithms Problems / Dynamic Programming / Minimum Deletions to Make String Balanced
+Algorithms Problems / Dynamic Programming / Stone Game
+Algorithms Problems / Dynamic Programming / Target Sum
+Algorithms Problems / Fenwick Tree / Distribute Elements Into Two Arrays II
+Algorithms Problems / Greedy / Max Difference You Can Get From Changing an Integer
+Algorithms Problems / Greedy / Maximum Length of Subarray With Positive Product
+Algorithms Problems / Math / Construct the Minimum Bitwise Array II
+Algorithms Problems / Math / Egg Drop With 2 Eggs and N Floors
+Algorithms Problems / Math / Find Number of Ways to Reach the K-th Stair
+Algorithms Problems / Math / Minimum Moves to Capture The Queen
+Algorithms Problems / Precomputation / Range Product Queries of Powers
+Algorithms Problems / Sliding Window / Find the Longest Equal Subarray
+Algorithms Problems / String / License Key Formatting
+Algorithms Problems / Traversal / Find the Number of Ways to Place People I
+Algorithms Problems / Traversal / Maximum Number of Operations With the Same Score I
+Algorithms Problems / Two Pointer / Boats to Save People
+Algorithms Problems / Two Pointer / Find the Lexicographically Largest String From the Box I
+Algorithms Problems / Two Pointer / Merge Sorted Array
+Algorithms Problems / SQL / Odd and Even Transactions
